requirements.txt
fastapi
uvicorn[standard]
requests
google-genai>=1.51.0
pymupdf
python-dotenv
httpx
